Remove dead GPIO/HAP helpers and fold hap/server types.
Drop unused channel_gpio, doctor path re-exports, and meta layer test;
merge types into handlers and tighten test-only visibility.
