+++ src/hap/server/mod.rs
@@ -4,7 +4,6 @@
 mod handlers;
 mod state;
 mod transport;
-mod types;
 
 use anyhow::Result;
 use std::net::SocketAddr;
@@ -91,89 +90,4 @@ where
         }
     }
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::handlers::{
-        build_event_body, characteristics_body, parse_characteristic_ids, write_statuses_body,
-    };
-    use crate::hap::runtime::{
-        CharacteristicEvent, CharacteristicId, CharacteristicRead, CharacteristicWriteStatus,
-        HapStatus, Subscriptions,
-    };
-    use serde_json::json;
-
-    #[test]
-    fn event_body_filters_to_subscribed_characteristics() {
-        let event = CharacteristicEvent {
-            id: CharacteristicId::new(2, 9),
-            value: json!(100),
-        };
-        let mut subs = Subscriptions::default();
-        assert!(build_event_body(std::slice::from_ref(&event), &subs).is_none());
-
-        subs.insert(event.id);
-        let body = build_event_body(&[event], &subs).unwrap();
-        let parsed: serde_json::Value = serde_json::from_slice(&body).unwrap();
-        assert_eq!(parsed["characteristics"][0]["aid"], 2);
-        assert_eq!(parsed["characteristics"][0]["iid"], 9);
-        assert_eq!(parsed["characteristics"][0]["value"], 100);
-    }
-
-    #[test]
-    fn parses_characteristic_ids() {
-        let ids = parse_characteristic_ids("2.9,3.10,bad");
-
-        assert_eq!(ids[0], CharacteristicId::new(2, 9));
-        assert_eq!(ids[1], CharacteristicId::new(3, 10));
-        assert_eq!(ids[2], CharacteristicId::new(0, 0));
-    }
-
-    #[test]
-    fn characteristics_body_uses_status_for_read_errors() {
-        let body = characteristics_body(vec![CharacteristicRead::error(
-            CharacteristicId::new(2, 99),
-            HapStatus::ResourceDoesNotExist,
-        )]);
-        let parsed: serde_json::Value = serde_json::from_slice(&body).unwrap();
-
-        assert_eq!(parsed["characteristics"][0]["aid"], 2);
-        assert_eq!(parsed["characteristics"][0]["iid"], 99);
-        assert_eq!(
-            parsed["characteristics"][0]["status"],
-            HapStatus::ResourceDoesNotExist.code()
-        );
-        assert!(parsed["characteristics"][0].get("value").is_none());
-    }
-
-    #[test]
-    fn write_statuses_body_reports_per_characteristic_status() {
-        let body = write_statuses_body(vec![CharacteristicWriteStatus::error(
-            CharacteristicId::new(2, 9),
-            HapStatus::ReadOnly,
-        )]);
-        let parsed: serde_json::Value = serde_json::from_slice(&body).unwrap();
-
-        assert_eq!(parsed["characteristics"][0]["aid"], 2);
-        assert_eq!(parsed["characteristics"][0]["iid"], 9);
-        assert_eq!(
-            parsed["characteristics"][0]["status"],
-            HapStatus::ReadOnly.code()
-        );
-    }
-
-    #[test]
-    fn server_runtime_layer_does_not_import_somfy_modules() {
-        let root = concat!(env!("CARGO_MANIFEST_DIR"), "/src/hap/server");
-        for name in ["mod.rs", "handlers.rs", "transport.rs"] {
-            let source = std::fs::read_to_string(format!("{root}/{name}")).unwrap();
-            assert!(!source.contains(concat!("crate::", "gpio")), "{name}");
-            assert!(!source.contains(concat!("crate::", "remote")), "{name}");
-            assert!(
-                !source.contains(concat!("crate::server::", "AppState")),
-                "{name}"
-            );
-        }
-    }
 }
